This PR was squashed before being merged into the main branch.

Discussion
----------

 Run phpstan on example

| Q             | A
| ------------- | ---
| Bug fix?      | no
| New feature?  | no <!-- please update src/**/CHANGELOG.md files -->
| Docs?         | no <!-- required for new features -->
| Issues        | Fix #520 <!-- prefix each issue number with "Fix #", no need to create an issue if none exist, explain below instead -->
| License       | MIT

Closes https://github.com/symfony/ai/issues/520

Commits
-------

1e436e55  Run phpstan on example

+++ examples/rag/chromadb.php
@@ -34,7 +34,7 @@ require_once dirname(__DIR__).'/bootstrap.php';
 $store = new Store(
     (new Codewithkyrian\ChromaDB\Factory())
         ->withHost(env('CHROMADB_HOST'))
-        ->withPort(env('CHROMADB_PORT'))
+        ->withPort((int) env('CHROMADB_PORT'))
         ->connect(),
     'movies',
 );
